move to utils file
