+++ packages/api-auth/index.ts
@@ -1,6 +1,7 @@
 import { createRemoteJWKSet, jwtVerify } from "jose";
 import { FastifyInstance, FastifyRequest, FastifyReply } from "fastify";
 import fp from "fastify-plugin";
+import { getMapFromScope, validatePermission } from "./utils.js";
 
 type ExtractedUserData = {
   userId: string;
@@ -8,8 +9,6 @@ type ExtractedUserData = {
 };
 
 type MatchConfig = { method: "AND" | "OR" };
-
-export type ScopeMap = Map<string, ScopeMap | boolean>;
 
 declare module "fastify" {
   interface FastifyRequest {
@@ -39,56 +38,6 @@ const decodeLogtoToken = async (
     issuer: config.oidcEndpoint,
   });
   return payload;
-};
-
-export const getMapFromScope = (scope: string) => {
-  const scopes = scope.split(" ");
-
-  return scopes.reduce<ScopeMap>((acc, scope) => {
-    const subScope = scope.split(":");
-    let current: ScopeMap | boolean | undefined = acc;
-
-    for (let i = 0; i < subScope.length; i++) {
-      const part = subScope[i];
-
-      if (current === true) break;
-
-      if (current instanceof Map) {
-        if (subScope[i + 1] === "*") {
-          current.set(part, true);
-          break;
-        }
-
-        if (i === subScope.length - 1) {
-          current.set(part, true);
-        } else if (!current.get(part)) {
-          current.set(part, new Map());
-        }
-
-        current = current.get(part);
-      }
-    }
-
-    return acc;
-  }, new Map());
-};
-
-export const validatePermission = (permission: string, scope: ScopeMap) => {
-  const parts = permission.split(":");
-
-  let current: ScopeMap | boolean | undefined = scope;
-
-  for (let i = 0; i <= parts.length; i++) {
-    const part = parts[i];
-
-    if (current === true) return true;
-
-    if (current instanceof Map) {
-      current = current.get(part);
-    } else {
-      return false;
-    }
-  }
 };
 
 export const checkPermissions = async (
